forces app to trust the 1st proxy in-front of it
for example we have app launched in docker container in front-of
it we have nginx, and nginx is covered by cloudflare.

app.set('trust proxy', 1);: Informs Express to trust the first
proxy in front of it. This setting adjusts how Express
calculates req.ip and other connection properties. Since
your app is behind Nginx (which is behind Cloudflare), setting
it to 1 tells Express to use the client IP provided by Nginx.